+++ mitra_cli/src/cli.rs
@@ -63,6 +63,7 @@ use mitra_models::{
         find_empty_profiles,
         find_unreachable,
         get_profile_by_id,
+        set_profile_identity_key,
     },
     subscriptions::queries::{
         reset_subscriptions,
@@ -139,6 +140,7 @@ pub enum SubCommand {
     ListUsers(ListUsers),
     SetPassword(SetPassword),
     SetRole(SetRole),
+    EnableFepEf61(EnableFepEf61),
     FetchActor(FetchActor),
     ReadOutbox(ReadOutbox),
     FetchReplies(FetchReplies),
@@ -374,6 +376,27 @@ impl SetRole {
         let role = role_from_str(&self.role)?;
         set_user_role(db_client, &profile.id, role).await?;
         println!("role changed");
+        Ok(())
+    }
+}
+
+#[derive(Parser)]
+pub struct EnableFepEf61 {
+    username: String,
+}
+
+impl EnableFepEf61 {
+    pub async fn execute(
+        &self,
+        db_client: &mut impl DatabaseClient,
+    ) -> Result<(), Error> {
+        let user = get_user_by_name(db_client, &self.username).await?;
+        set_profile_identity_key(
+            db_client,
+            user.id,
+            user.ed25519_private_key,
+        ).await?;
+        println!("FEP-ef61 enabled for actor");
         Ok(())
     }
 }