+++ mitra_models/migrations/V0077__actor_profile__identity_key.sql
@@ -1,0 +1,1 @@
+ALTER TABLE actor_profile ADD COLUMN identity_key VARCHAR(100) UNIQUE;