+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Implemented FEP-ef61 resolver endpoint.
 - Use actor ID as handle if acct URI is not known.
+- Added `enable-fep-ef61` command.
 
 ### Changed
 
@@ -19,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Don't rely on acct comparsion when verifying activity signature.
 - Use actor ID instead of webfinger address in logs.
 - Use actor ID as primary identifier instead of webfinger address.
+- Don't publish FEP-ef61 representation if user didn't enable FEP-ef61.
 
 ### Fixed